use evm compatible accounts, and fix properties for chain spec (#326)
* use evm compatible accounts, and fix properties for chain spec

* doc changes

* bump node and docs versions

* impl_runtime_apis need to be in the same file with construct runtime for the moment

* account type fix

* zombienet script update

* slot duration fix

---------
