+++ evm-template/node/Cargo.toml
@@ -7,7 +7,7 @@ edition = { workspace = true }
 license = { workspace = true }
 publish = false
 repository = { workspace = true }
-version = "1.0.0"
+version = "2.0.0"
 
 [dependencies]
 clap = { workspace = true }
